[DEV] Update axivion compiler config to new format
Use `migrate compiler` command to update the compiler config to the new format as described in the migration guide of Axivion 7.3.*. The resulting `compiler_config.json` is then moved to the desired folder and replaces the old configuration files.
NOTE: Since we are using two `axivion_config.json` files (common and project specific) we have to set the `BAUHAUS_CONFIG` accordingly before doing the migration (cf. `start_analysis.sh`).